requirements.txt
python-telegram-bot[webhooks]==22.7
anthropic==0.86.0
pydantic==2.12.5
pydantic-settings==2.13.1
aiosqlite==0.22.1
kerykeion==5.12.6
pyswisseph==2.10.3.2
geopy==2.4.1
timezonefinder==8.2.1
Pillow==12.1.1
loguru==0.7.3
python-dotenv==1.2.2
pytest==9.0.2
pytest-asyncio==1.3.0
telethon==1.39.0
